Mise a jour de la class Author service avec orElseThrow , ajout de responseEntity pour le controller

+++ backDev/bookstore/src/main/java/com/library/bookstore/repository/AuthorRepository.java
@@ -3,5 +3,6 @@ package com.library.bookstore.repository;
 import com.library.bookstore.entity.Author;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AuthorRepository extends JpaRepository<Author, Integer> {
+public interface AuthorRepository extends JpaRepository<Author, Long> {
+
 }
